+++ src/test/java/com/github/mahendranv/AnchorParserTest.java
@@ -17,6 +17,7 @@ public class AnchorParserTest {
         assertEquals(108, result.getChannel().getItems().size());
         assertTrue(channel.isExplicit());
         assertTrue(channel.getItems().get(0).isExplicit());
+        assertNotNull(channel.getItems().get(0).getEnclosure().getUrl());
         assertEquals("Comedy", channel.getCategory());
     }
 
@@ -26,6 +27,7 @@ public class AnchorParserTest {
         Channel channel = assertMandatoryFields(result);
         assertFalse(channel.isExplicit());
         assertFalse(channel.getItems().get(0).isExplicit());
+        assertNotNull(channel.getItems().get(0).getEnclosure().getUrl());
         assertEquals(242, channel.getItems().size());
         assertEquals("Technology", channel.getCategory());
     }